# public/openapi.yaml
openapi: 3.0.3
info:
  title: 'Laravel Swagger API Documentation'
  description: 'Backend API for the salon SaaS platform (Laravel + PostgreSQL) with JWT auth and tenant-scoped operations.'
  version: 1.0.0
servers:
  -
    url: https://admin.vynhq.com
    description: Production
  -
    url: http://127.0.0.1:8000
    description: Local (php artisan serve)
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: 'Use `/api/auth/login` (or related auth endpoints) to obtain a JWT token. Include `X-Tenant` for tenant-scoped endpoints. Authentication is configured globally in the OpenAPI security scheme.'
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  -
    default: []
    bearerAuth: []
paths:
...
  /api/otp/verify:
    post:
      summary: 'Verify OTP'
      operationId: verifyOTP
      description: ''
      parameters:
        -
          in: header
          name: X-Tenant
          description: ''
          example: '1'
          schema:
            type: string
      responses:
        422:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  success: false
                  message: 'Validation failed'
                  errors:
                    identifier:
                      - 'The identifier field is required.'
                    code:
                      - 'The code field is required.'
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'Validation failed'
                  errors:
                    type: object
                    properties:
                      identifier:
                        type: array
                        example:
                          - 'The identifier field is required.'
                        items:
                          type: string
                      code:
                        type: array
                        example:
                          - 'The code field is required.'
                        items:
                          type: string
      tags:
        - Authentication
      security: []
...
  /api/auth/register/customer:
    post:
      summary: 'Register Customer'
      operationId: registerCustomer
      description: ''
      parameters:
        -
          in: header
          name: X-Tenant
          description: ''
          example: '1'
          schema:
            type: string
      responses:
        422:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  success: false
                  message: 'Validation failed'
                  errors:
                    email:
                      - 'The email field is required.'
                    password:
                      - 'The password field is required.'
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'Validation failed'
                  errors:
                    type: object
                    properties:
                      email:
                        type: array
                        example:
                          - 'The email field is required.'
                        items:
                          type: string
                      password:
                        type: array
                        example:
                          - 'The password field is required.'
                        items:
                          type: string
      tags:
        - Authentication
      security: []
  /api/auth/register/salon-owner:
    post:
      summary: 'Register Salon Owner'
      operationId: registerSalonOwner
      description: ''
      parameters:
        -
          in: header
          name: X-Tenant
          description: ''
          example: '1'
          schema:
            type: string
      responses:
        422:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  success: false
                  message: 'Validation failed'
                  errors:
                    salon_name:
                      - 'The salon name field is required.'
                    email:
                      - 'The email field is required.'
                    password:
                      - 'The password field is required.'
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'Validation failed'
                  errors:
                    type: object
                    properties:
                      salon_name:
                        type: array
                        example:
                          - 'The salon name field is required.'
                        items:
                          type: string
                      email:
                        type: array
                        example:
                          - 'The email field is required.'
                        items:
                          type: string
                      password:
                        type: array
                        example:
                          - 'The password field is required.'
                        items:
                          type: string
      tags:
        - Authentication
      security: []
...
  /api/auth/verify-otp:
    post:
      summary: 'Verify OTP'
      operationId: verifyOTP
      description: ''
      parameters:
        -
          in: header
          name: X-Tenant
          description: ''
          example: '1'
          schema:
            type: string
      responses:
        422:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  success: false
                  message: 'Validation failed'
                  errors:
                    identifier:
                      - 'The identifier field is required.'
                    code:
                      - 'The code field is required.'
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'Validation failed'
                  errors:
                    type: object
                    properties:
                      identifier:
                        type: array
                        example:
                          - 'The identifier field is required.'
                        items:
                          type: string
                      code:
                        type: array
                        example:
                          - 'The code field is required.'
                        items:
                          type: string
      tags:
        - Authentication
      security: []
  /api/auth/google:
    post:
      summary: 'Google Auth'
      operationId: googleAuth
      description: ''
      parameters:
        -
          in: header
          name: X-Tenant
          description: ''
          example: '1'
          schema:
            type: string
      responses:
        422:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  success: false
                  message: 'Validation failed'
                  errors:
                    credential:
                      - 'The credential field is required.'
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'Validation failed'
                  errors:
                    type: object
                    properties:
                      credential:
                        type: array
                        example:
                          - 'The credential field is required.'
                        items:
                          type: string
      tags:
        - Authentication
      security: []
...